Week2 STEP 3 수정사항 반영
